Implemented the set() functions and operators for FactBase
All the set() functions and operators have been added. For the operators that
compare to other objects you can use either another FactBase or some other
collection. If you use another collection then a temporary FactBase will be
created, which will also ensure that an error is raised if an object is not a
predicate instance.

Not sure if this has been worth it, maybe would have been better to require
users to convert to/from a standard set for any set operations. We will see.