Extend clorm.clingo.Control.assign_external() to allow collections
The assign_external() and release_external() member functions can now take an
atom represented by a clingo Symbol or literal (int) or a clorm Predicate, or it
can take a collection of atoms.

Also fixed the implementation in clorm/clingo.py to check for Iterable whenever
we could be dealing with an collection.