Tweak calculation of automatically generated predicate name.
If first letter is lower-case then do nothing. Otherwise, replace any sequence of
upper-case only characters that occur at the beginning of the string or
immediately after an underscore with lower-case equivalents. The sequence of
upper-case characters can include non-alphabetic characters (eg., numbers) and
this will still be treated as a single sequence of upper-case characters.

The above covers basic naming conventions: camel-case, snake-case, and acronyms.