More error checks for Select/Delete 'where' 'order_by' expressions
Added more checking for expressions that don't refer to fields of the predicate
type associated with the Select/Delete query. This will improve debugging as it
will generate the error when the query is declared rather than waiting for the
query to be executed to generate the error. Note: for 'where' clauses it is not
possible to catch all bad expressions, since an arbitrary functor can be
supplied rather than a comparator expression.