Weakened the strict instance type requirement for WrapperMetaClass
It is now possible to pass an object to a defined wrapper class that is not
actually of that wrapped class type. The motivation is to be able to wrap clingo
theory extensions, such as clingcon and clingo-lp, with a clingo.Control
interface and then be able to pass such an object to clorm.clingo.Control in the
control_ parameter.