More checks and flexibility for the clorm.clingo wrapper
Added checks that an explicitly declared wrapped Control object has 'solve' and
'ground' functions, and similarly that a explicitly wrapped Model object has a
'symbols' function. Also made the Model.facts() function more flexible so it
passes through any unrecognised arguments to the underlying symbols() function
call (making it easier to implement a wrapper of a wrapper - which is what I
want to do with clingo theory propagators).