Fixed query sort order assignment bugs
Fixed how the order_by components were distributed to each JoinQueryPlan and
also assigned within an individual JoinQueryPlan. It is now clearer when an
OrderByBlock is assigned to a prejoin component and a postjoin component.

Note: For a JoinQueryPlan with an OrderByBlock there should be no OrderByBlock
in any previous JoinQueryPlan, because any previous orderings will be
overwritten by this one.