Fixed logical bug in setting up query sort ordering
Previous implementation only worked for a special case. So removed any of the
prejoin sorts and only have a single sort point within the query. While this is
not optimal it is probably good enough for the size of problems that we
encounter. Still, I would like to improve it at some point.