Move noclingo and removed noclingo patching.
The noclingo module has been moved into the orm sub-package. Trying to remove
the need to import clingo with a noclingo patch isn't useful. Instead the goal
for v2 is to only create clingo.Symbol objects when explicitly accessing the
raw() member function, for example as part of asserting to a clingo.Control
object. So, noclingo will be used as a substitute in special cases such as a
where the user wants a RawField but doesn't need to push it directory to
clingo.Control.