Use only the field values to compute hash facts
Used to use the raw clingo Symbol object, but changed to only use the fields for
better de-coupling from clingo. This also forces all field assignments to be for
hashable objects. While this doesn't guarantee that the object is immutable at
least it forces users to be more aware that field objects should be immutable.