Partial Predicate instance and FactBase pickling
Class instances to be pickled must be declared globally. This means that
predicate sub-classes that are declared dymanically don't work; for example
having a tuple in a specification of a predicate field or using
simple_predicate(). Need to see if there is some workaround; eg. to export the
sub-class definitions globally.