Bug in unification that failed for unary predicates
Because a fact behaves like a collection of elements (like a list or tuple)
therefore a unary predicate behaves like a collection with no elements. This
means that testing the booleaness of a unary fact will return False, in the same
way that testing an empty tuple returns False.

So needed to change the test for successful unification to a "is None" test
rather than whether the fact returns True/False.