Added a query 'ordered()' flag as a shorthand for the natural order
For a set of roots P1,...PN the 'ordered()' flag is a short hand for specifying
'order_by(asc(P1),...,asc(PN))'. This can be useful to remove any non
determinism in the output.  However, it is heavy handed so still work looking
into the query specification to see if we can remove non-determinism when
running the same program at different times.