Moved helper functions into clorm namespace (from clorm.clingo)
Some refactoring to move the `control_add_facts()` and
`symbolic_atoms_to_facts()` functions from the `clorm.clingo` namespace to the
main `clorm` namespace. `clorm.clingo` should only be for the convenience
`clorm.clingo.Control` wrapper and not for general helper functions that you
would use when interacting with a `clingo.Control` object.