Added readthedocs configuration to fix missing typing_extensions
The docs are not being built correctly because it is failing to import the
typing_extensions module. I can't find good references to what is going
on (whether for some version of Python 3.5+ typing_extensions is now part of
the standard and for others you need to pip install it. So trying to use Python
3.9 to build the docs.