Custom creation of Predicate init function by calling exec()
For each Predicate sub-class a custom init function is generated as a string
and then turned into Python bytecode by calling exec(). This creates a much
leaner and faster initialisation, since as much work as possible is done when
defining the init function and not have to be built into the logic of the init
function itself.