Use Predicate template for __bool__ and __len__
Using the templating for Predicate sub-class __bool__() and __len__() because
their values can be worked out when the sub-class is created.

Slight change of behaviour for __bool__() that it now only behaves like a tuple
if the predicate is declared as a tuple. So returns False only if it is an
empty tuple, otherwise returns true. This seems like more appropriate
behaviour.