Getting FactBase to work with new Predicate semantics
Previously, Predicate comparison operators were overloaded to deal with
standard Python tuples. This made using standard tuples in queries
easy. However, iith the change of semantics we now have to detect when a python
tuple is specified in a query that it needs to be explicitly converted to the
appropriate clorm complex tuple type.