Fixed big in predicate order for queries
The bug occured when passing a function/lambda as the "select" clause for a query. It wasn't
properly detecting that the input to the function/lambda needed to be the fact objects in the
order specified by the query roots (and not the internal ordering created by the query plan).